Fix missing header when same class name is used in two modules
When the same class name was used in two or more modules which were
added to the report in sequence, only the first class name was
displayed. It happened because the plugin tracked differences in
Result.header in each test result and only the class name (without its
module) was set in it.